add ParseFileAsFileDescriptors test helper
