+++ adapter/presenter/cli_test.go
@@ -14,12 +14,10 @@ func TestCLIPresenter(t *testing.T) {
 	presenter := NewJSONCLIPresenter()
 
 	t.Run("Call", func(t *testing.T) {
-		env := testhelper.SetupEnv(t, "helloworld.proto", "helloworld", "Greeter")
+		descs := testhelper.ReadProtoAsFileDescriptors(t, "helloworld.proto")
+		msg := testhelper.FindMessage(t, "HelloRequest", descs)
 
-		msg, err := env.Message("HelloRequest")
-		require.NoError(t, err)
-
-		dmsg := dynamic.NewMessage(msg.Desc)
+		dmsg := dynamic.NewMessage(msg)
 		dmsg.SetField(dmsg.FindFieldDescriptorByName("name"), "makise")
 		dmsg.SetField(dmsg.FindFieldDescriptorByName("message"), "kurisu")
