+++ entity/message.go
@@ -14,14 +14,13 @@ type Message struct {
 	NestedMessages Messages
 	NestedEnums    []*Enum
 
-	// TODO: I wanna change to unexported
-	Desc      *desc.MessageDescriptor
+	desc      *desc.MessageDescriptor
 	fieldDesc *desc.FieldDescriptor // fieldDesc is nil if this message is not used as a field
 }
 
 func newMessage(m *desc.MessageDescriptor) *Message {
 	msg := Message{
-		Desc: m,
+		desc: m,
 	}
 
 	msgs := make(Messages, len(m.GetNestedMessageTypes()))
@@ -69,7 +68,7 @@ func (m *Message) Name() string {
 	if m.fieldDesc != nil {
 		return m.fieldDesc.GetName()
 	}
-	return m.Desc.GetName()
+	return m.desc.GetName()
 }
 
 func (m *Message) Type() string {
